Refactor: adds functionality to disable use button in voucher when minimum transaction amount condition not meet

+++ pulsa/src/main/java/com/example/pulsa/ui/plans/mapper/PlansItemResponseDomainToUiMapper.kt
@@ -10,7 +10,7 @@ class PlansItemResponseDomainToUiMapper @Inject constructor() :
     override fun map(input: PlansItemResponseDomainModel) = PlansItemResponseUiModel(
         productCode = input.productCode,
         productName = input.label,
-        rechargeAmount = input.price.toBigDecimal(),
-        talkTimeAmount = input.nominal.toBigDecimal(),
+        rechargeAmount = input.price,
+        talkTimeAmount = input.nominal,
     )
 }
